pyproject.toml
[tool.poetry]
name = "alphacodium-dspy"
version = "0.1.0"
description = "AlphaCodium implementation using DSPy"
authors = ["John Pertoft <john.pertoft@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
dspy-ai = "^2.1.10"
datasets = "^2.14.6"
pysnooper = "^1.2.0"
dynaconf = "^3.2.4"
loguru = "^0.7.2"
typer = "^0.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
no_implicit_optional = true

[tool.isort]
profile = "black"
force_single_line = true
